+++ docs/CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- None
+- Grouped related settings dialog controls for clearer scanning.
 
 ### Deprecated